+++ src/utils.c
@@ -115,6 +115,32 @@ SEXP rnng_sleep(SEXP msec) {
 
 }
 
+SEXP rnng_random(SEXP n) {
+
+  SEXP vec;
+  R_xlen_t vlen;
+  switch (TYPEOF(n)) {
+  case INTSXP:
+  case LGLSXP:
+    vlen = (R_xlen_t) INTEGER(n)[0];
+    break;
+  case REALSXP:
+    vlen = (R_xlen_t) Rf_asInteger(n);
+    break;
+  default:
+      Rf_error("'n' must be integer or coercible to integer");
+  }
+
+  vec = Rf_allocVector(RAWSXP, vlen * 4);
+  void *pvec = STDVEC_DATAPTR(vec);
+  for (R_xlen_t i = 0; i < vlen; i++, pvec += 4) {
+    u_int32_t rand = nng_random();
+    memcpy(pvec, &rand, 4);
+  }
+  return vec;
+
+}
+
 SEXP rnng_url_parse(SEXP url) {
 
   const char *up = CHAR(STRING_ELT(url, 0));